move zoekt tests from textsearch_test to zoekt_test
